Update blocked users in real time
